requirements.txt
pyspark
pyyaml
# JDBC drivers for different databases
# postgresql-jdbc (included with pyspark)
# mysql-connector-java (for MySQL)
# ojdbc8 (for Oracle)
# mssql-jdbc (for SQL Server) 